refactor(eve): split workflow finalization
